Exclude unset addresses in returned map
If an address has not been set in global state, it should be omitted from the
map of returned address, data pairs returned by State.Get().

Currently, there is no way to differentiate between a value that is set to no
bytes, and an unset address, so this treats both as an unset address.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>